% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{aggregateIncidence}
\alias{aggregateIncidence}
\title{Run Aggregation of Incidence Query}
\usage{
aggregateIncidence(connectionDetails, cdmSourceName, cdmDatabaseSchema,
  resultsDatabaseSchema, cohortDefinitionId, cohortTableName = "cohort")
}
\value{
none
}
\description{
Run Aggregation of Incidence Query
}
\details{
Generates a cohort's overall incidence in person-time for a CDM
}

