% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{characterize}
\alias{characterize}
\title{Run Phenotype Characterization Query}
\usage{
characterize(connectionDetails, cdmDatabaseSchema, resultsDatabaseSchema,
  cohortDefinitionId)
}
\value{
none
}
\description{
Run Phenotype Characterization Query
}
\details{
Generates table of cohort measures, stratified
}

