% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotIncidenceTrellis}
\alias{plotIncidenceTrellis}
\title{plotIncidenceTrellis
Plot prevalence stratified by age, gender}
\usage{
plotIncidenceTrellis(result)
}
\arguments{
\item{result}{The data frame with your stratified incidence data}
}
\value{
none
}
\details{
Plot prevalence stratified by age, gender in trellis
}

