% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotPrevalanceTrellis}
\alias{plotPrevalanceTrellis}
\title{Plot prevalence stratified by age, gender}
\usage{
plotPrevalanceTrellis(result)
}
\value{
none
}
\description{
Plot prevalence stratified by age, gender
}
\details{
Plot prevalence stratified by age, gender in trellis
}

