% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runDrugOutcomeSummary}
\alias{runDrugOutcomeSummary}
\title{Run Drug Outcome Summary}
\usage{
runDrugOutcomeSummary(connectionDetails, cdmSourceName, cdmDatabaseSchema,
  resultsDatabaseSchema, cohortDefinitionId, cohortTableName = "cohort")
}
\value{
none
}
\description{
Run Drug Outcome Summary
}
\details{
Runs Drug Outcome Summary query
}

